Add files via upload
There are two objectives this thesis aims to accomplish. The first goal is to provide a methodology in surveying and understanding the global governing structures of the Internet System and cyberspace. The second objective of this paper is to provide an overview of the current state assessment of the governance architecture, and infer from its design what the principles of these policies yield and as well as their intention.

This paper asserts a methodology of appreciating the whole of the complex system, rather than a skewed focus of the sum of its parts. In viewing the whole scope, can theories of international relation evolve and be better adaptive to the shifting changes in today’s brave new world. Through this methodology, the findings of this paper suggests the current existing governance architecture yields design principles conducive to emergent characteristics conducive to creating a space and practice for innovation development, as well as room for self-adaptation.